requirements.txt
opencv-python
gymnasium
tetris-gymnasium
torch
torchvision
pillow
numpy
